requirements.txt
Flask==3.1.2
Werkzeug==3.1.4
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
cryptography==46.0.3
gunicorn==23.0.0
qrcode==8.2
Pillow==12.0.0
pytest==9.0.1